move code in 'etl' subpackage
