+++ mcs_formalizations/etl/_pipeline.py
