+++ mcs_formalizations/etl/nop_extractor.py
@@ -1,0 +1,6 @@
+from mcs_formalizations._extractor import _Extractor
+
+
+class NopExtractor(_Extractor):
+    def extract(self, **kwds):
+        return {}
