+++ mcs_formalizations/etl/pipelines/automobile/automobile_object_pipeline.py
