+++ mcs_formalizations/etl/pipelines/automobile/automobile_object_transformer.py
