+++ mcs_formalizations/etl/pipelines/parking_garage/parking_garage_object_pipeline.py
