--- mcs_formalizations/nop_extractor.py
@@ -1,6 +1,0 @@
-from mcs_formalizations._extractor import _Extractor
-
-
-class NopExtractor(_Extractor):
-    def extract(self, **kwds):
-        return {}
